% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brcaData}
\alias{brcaData}
\title{breast tumor data}
\format{
A data frame with 54675 rows and 47 variables:
}
\usage{
brcaData
}
\description{
data from GEO: https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE3744
}
\keyword{datasets}
